Refectoring

+++ BookStore.DataAccess/Repositories/Interfaces/IBaseEFRepository.cs
@@ -9,7 +9,7 @@ namespace BookStore.DataAccess.Repositories.Interfaces
     {
         Task<IEnumerable<TEntity>> GetAsync();
         IEnumerable<TEntity> Get(Func<TEntity, bool> predicate);
-        Task<TEntity> FindByIdAsync(int id);
+        Task<TEntity> FindByIdAsync(long id);
         Task CreateAsync(TEntity item);
         Task UpdateAsync(TEntity item);
         Task RemoveAsync(TEntity item);
